The actual Notebook That I am working in as I work
How many records are in the file?
What are the sequence lengths?
What is the length of the longest ORF?
For a given sequence identifier, what is the longest ORF contained in the sequence represented by that identifier?
What is the starting position of the longest ORF in the sequence that contains it? 
Given a length n, your program should be able to identify all repeats of length n in all sequences in the FASTA file. Your program should also determine how many times each repeat occurs in the file, and which is the most frequent repeat of a given length.